+++ js/data.js
@@ -1,0 +1,11 @@
+const photos  = [];
+
+const messages = [
+  'Всё отлично!', 'В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.', 'В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+const descriptions = ['У меня все хорошо!', 'Мацарайва и ультравайоленс.', 'Подружки топ.', 'Пилоты выпустили новый альбом.', 'Лана дель рей это икона.', 'Люблю цветы!', 'Мой котик иногда глупый, но я его люблю.',];
+
+const names = ['Ярослава', 'Мирожанна', 'Алиса', 'Нацу', 'Люси', 'Эрза', 'Леви', 'Слоумо', 'Венди'];
+
+export {photos, descriptions, messages, names};
