+++ js/main.js
@@ -1,33 +1,25 @@
-const messages = [
-  'Всё отлично!', 'В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.', 'В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+import {photos, descriptions, messages, names} from '/.data.js';
 
-const descriptions = ['У меня все хорошо!', 'Мацарайва и ультравайоленс.', 'Подружки топ.', 'Пилоты выпустили новый альбом.', 'Лана дель рей это икона.', 'Люблю цветы!', 'Мой котик иногда глупый, но я его люблю.',];
-
-const names = ['Ярослава', 'Мирожанна', 'Алиса', 'Нацу', 'Люси', 'Эрза', 'Леви', 'Слоумо', 'Венди'];
+import {randomInteger} from '/.util.js';
 
 const COUNT_PHOTO =  25;
 
-const photos  = [];
-
-const comments = {
+const Comments = {
   MIN: 0,
   MAX: 30
 };
-const ids = {
+const Ids = {
   MIN: 1,
   MAX: 100
 };
 
-const likes = {
+const Likes = {
   MIN: 15,
   MAX: 200
 };
 
-const randomInteger = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
-
 const addComment = () => ({
-  id: randomInteger(ids.MIN, ids.MAX),
+  id: randomInteger(Ids.MIN, Ids.MAX),
   avatar: `img/avatar-${randomInteger(1, 6)}}.svg`,
   message: messages[randomInteger(0, messages.length - 1)],
   name: names[randomInteger(0, names.length - 1)]
@@ -35,11 +27,11 @@ const addComment = () => ({
 
 const generateComments = () => {
   const commentsArray = [];
-  const commentsCount = randomInteger(comments.MIN, comments.MAX);
+  const commentsCount = randomInteger(Comments.MIN, Comments.MAX);
   for (let i = 0; i < commentsCount; i++) {
     commentsArray.push(addComment());
   }
-  return comments;
+  return Comments;
 };
 
 
@@ -47,7 +39,7 @@ const addPhoto = (index) => ({
   id: index + 1,
   url: `photos/${index + 1}.jpg`,
   description: descriptions[randomInteger(0, descriptions.length-1)],
-  likes: randomInteger(likes.MIN, likes.MAX),
+  likes: randomInteger(Likes.MIN, Likes.MAX),
   comments: generateComments()
 });
