+++ js/util.js
@@ -1,0 +1,3 @@
+const randomInteger = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+export {randomInteger};
